+++ lib/coolkit/hash/sash.rb
@@ -1,0 +1,3 @@
+module Coolkit
+
+end