+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 0.5.0 (Unreleased)
 
-## 0.4.0 (2019-12-21) [Unreleased]
+### Added
+
+- `Coolkit.full_extension`
+- `Coolkit.yaml2json` for converting YAML files to JSON. Handy for writing JSON schema files in YAML.
+- Assorted methods for finding duplicates in string arrays.
+
+## 0.4.0 (no release)
 
 ### Added